+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Hoodoo v2.x
 
+## 2.3.0 (2018-02-22)
+
+* Maintenance `bundle update`.
+* Introduces environment variable [`HOODOO_CLOCK_DRIFT_TOLERANCE`](http://loyaltynz.github.io/hoodoo/guides_1000_env_vars.html#hoodoo_clock_drift_tolerance) and related method [`Hoodoo::Utilities::is_in_future?`](https://cdn.rawgit.com/LoyaltyNZ/hoodoo/master/docs/rdoc/classes/Hoodoo/Utilities.html#method-c-is_in_future-3F).
+* Internal caching for a few more environment variables, yielding a small performance improvement.
+
 ## 2.2.3 (2018-02-19)
 
 General maintenance update and test matrix support for Ruby 2.5.0. No core function changes, fixes or new features.